requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.7.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
Flask==2.2.3
frozenlist==1.3.3
GDAL==3.6.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
openai==0.27.7
OpenCC==1.1.6
protobuf==4.21.12
pydantic==1.10.7
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
requests==2.30.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
text-unidecode==1.3
tqdm==4.65.0
typing_extensions==4.5.0
tzlocal==2.1
urllib3==2.0.2
uvicorn==0.21.1
Werkzeug==2.2.3
yarl==1.9.2
